refactor(quadraui): move build_shell_adapter to shell_adapter.rs, dedupe across backends
tui/shell_runner.rs factored AppShell construction into build_shell_adapter
"so the live runner and test driver can't drift"; gtk/shell_runner.rs never
adopted that factoring and re-inlined the same 45 lines (same builder-call
order, same magic numbers). Move the function (and its #547 regression
tests) into the already backend-neutral shell_adapter.rs, the single
canonical home for shell construction. Both shell_runner.rs modules are now
3-line wrappers: build adapter -> hand to the backend's run(). driver_with_shell
in both tui/testing.rs and gtk/testing.rs now call the same shared fn.

No pub API changes (build_shell_adapter was always pub(crate)); run_with_shell
signatures are unchanged, so no downstream impact for coord-tui or vimcode.

Unblocks #465: macOS's remaining scope is now just wiring a
macos::run_with_shell that calls this shared builder.

Closes #497.
